+++ sym/tests/aes_dec_tester.c
@@ -90,9 +90,9 @@ static int test_decrypt(const struct lc_sym *aes_impl, const char *name)
 	int ret;
 	LC_SYM_CTX_ON_STACK(aes, aes_impl);
 
-	printf("AES block ctx %s (%s implementation) len %lu\n", name,
+	printf("AES block ctx %s (%s implementation) len %u\n", name,
 	       aes_impl == lc_aes_c ? "C" : "accelerated",
-	       LC_SYM_CTX_SIZE(aes_impl));
+	       (unsigned int)LC_SYM_CTX_SIZE(aes_impl));
 
 	memcpy(in2, in256, sizeof(in256));
 	ret = test_decrypt_one(aes, key256, sizeof(key256), in2);
